rename file with kebab-case
